separated out iterator over different types of shard readers
